requirements.txt
Django==6.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==12.1.0
reportlab==4.2.5
python-dateutil==2.9.0
